requirements1.txt
fastapi==0.104.1
uvicorn==0.24.0
tensorflow==2.19.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.3
scikit-learn==1.3.2
python-multipart==0.0.6
